refactor(gateway-interceptors): use reflected protobuf codec
