+++ Cargo.toml
@@ -22,6 +22,7 @@ tonic-prost = "0.14"
 tonic-prost-build = "0.14"
 prost = "0.14"
 prost-types = "0.14"
+prost-reflect = { version = "0.16.5", features = ["serde"] }
 
 # HTTP server
 axum = { version = "0.8", features = ["ws"] }
