+++ crates/openshell-gateway-interceptors/Cargo.toml
@@ -13,11 +13,11 @@ repository.workspace = true
 [dependencies]
 openshell-core = { path = "../openshell-core", default-features = false }
 
-base64 = { workspace = true }
 hyper-util = { workspace = true, features = ["client", "http1", "http2", "tokio"] }
 json-patch = "1.4"
 metrics = { workspace = true }
 prost = { workspace = true }
+prost-reflect = { workspace = true }
 prost-types = { workspace = true }
 serde_json = { workspace = true }
 sha2 = { workspace = true }
